Fix: Replaced constexpr LU-tables generation by build time generation
Clang doesn't support very high amounts of compile time steps so we
replaced the generation process by a script generating the lookup tables
in a binary format, and embed it into the executable.

The reason behind is to avoid putting more runtime load and wait time
after starting the application.